+++ src/test/java/org/jenkinsci/plugins/codesonar/integration/conditions/NewWarningsIncreasedByPercentageConditionIT.java
@@ -65,9 +65,7 @@ public class NewWarningsIncreasedByPercentageConditionIT extends ConditionIntegr
         conditions.add(condition);
 
         CodeSonarPublisher codeSonarPublisher = new CodeSonarPublisher(
-                conditions, "http", VALID_HUB_ADDRESS.toString(), VALID_PROJECT_NAME, "",
-                this.VISIBILITY_FILTER
-        );
+                conditions, "http", VALID_HUB_ADDRESS.toString(), VALID_PROJECT_NAME, "", this.VISIBILITY_FILTER);
         codeSonarPublisher.setAnalysisService(mockedAnalysisService);
         codeSonarPublisher.setMetricsService(mockedMetricsService);
         codeSonarPublisher.setProceduresService(mockedProceduresService);